Backend/Summarization_Service/requirements.txt
django==5.1.2
djangorestframework==3.15.2
moviepy==1.0.3
whisper==1.0.0
transformers==4.25.2
pymongo==4.11
googletrans==4.0.0-rc1
python-dotenv==1.0.0